Kernel/Networking: Modify outgoing packet in place
This avoids creation of new buffers and unnecessary memcpys. The old
interface still exists for UDP but will be removed when I am less lazy.

From testing it does not appear to have any performance improvement but
this is most likely due to other bottlenecks as extra copies should
always be worse.