kernel: Add a way to parse numbers from C strings
After making this change I am now actually doubting if using
C strings is a good idea and maybe it should just always just the
string view library that userland makes use of.

But old code and the upcoming commits rely upon this so it is a change my
future less lazy self will do.